+++ infra/opentofu/talos-proxmox/providers.tf
@@ -1,0 +1,8 @@
+provider "proxmox" {
+  endpoint = var.proxmox_endpoint
+  username = var.proxmox_username
+  password = var.proxmox_password
+  insecure = true
+}
+
+provider "talos" {}